10 transfers access

+++ lessons/views.py
@@ -1,25 +1,27 @@
 import pytz
 from django.shortcuts import render, redirect
-from .forms import LessonRequestForm, StudentSignUpForm, LogInForm, BookLessonRequestForm, EditForm
+from .forms import LessonRequestForm, StudentSignUpForm, LogInForm, BookLessonRequestForm, EditForm, PasswordForm, UserForm
 from django.contrib.auth import authenticate, login, logout
 from django.contrib import messages
+from django.contrib.auth.hashers import check_password
 from django.contrib.auth.decorators import login_required
-from .models import Admin, LessonRequest, Lesson, Student, User, Invoice
-from .helpers import only_admins, only_students, get_next_given_day_of_week_after_date_given, find_next_available_invoice_number_for_student
+from .models import Admin, LessonRequest, Lesson, Student, User, Invoice, Transfer
+from .helpers import only_admins, only_students, get_next_given_day_of_week_after_date_given, find_next_available_invoice_number_for_student, login_prohibited, find_next_available_transfer_id
 from django.core.exceptions import ObjectDoesNotExist
+from django.utils import timezone
+
 
 import datetime
 
 # Create your views here.
+@login_prohibited
 def home(request):
     return render(request, 'home.html')
-
 @login_required
 @only_students
 def lessons_success(request):
     current_student_id = request.user.id
-    
-    lessons = Lesson.objects.filter(student_id=current_student_id) #get(student=current_user) # we filter this by the username and then pass it 
+    lessons = Lesson.objects.filter(student_id=current_student_id)
     return render(request, 'successful_lessons_list.html', {'lessons': lessons})
 
 @login_required
@@ -45,7 +47,7 @@ def lesson_request(request):
                 teacher=teacher
             )
             lessonRequest.save()
-            return redirect('lesson_request')
+            return redirect('student_home')
     else:
         form = LessonRequestForm()
     return render(request, 'lesson_request.html', {'form': form})
@@ -101,7 +103,6 @@ def book_lesson_request(request, request_id):
                 lesson.save()
                 new_date = new_date + tdelta
 
-            #TODO need to update this to set request to fulfilled and not delete it
             lesson_request.delete()
             return redirect('admin_requests')
     else:
@@ -117,7 +118,8 @@ def student_home(request):
 @only_admins
 def admin_home(request):
     return render(request, 'admin_home.html')
-    
+
+@login_prohibited
 def log_in(request):
     if request.method == 'POST':
         form = LogInForm(request.POST)
@@ -147,6 +149,7 @@ def log_out(request):
     logout(request)
     return redirect('home')
 
+@login_prohibited
 def student_sign_up(request):
     if request.method == 'POST':
         form = StudentSignUpForm(request.POST)
@@ -167,16 +170,52 @@ def admin_requests(request):
     return render(request, 'admin_lesson_list.html', {'data': lesson_request_data})
 
 @login_required
+@only_students
 def show_requests(request):
-    try:
-        user = request.user
-        lesson_requests = LessonRequest.objects.filter(author=user)
-    except ObjectDoesNotExist:
-        return redirect('home')
-    else:
-        return render(request, 'show_requests.html', {'user': user, 'lesson_requests': lesson_requests})
+    user = request.user
+    lesson_requests = LessonRequest.objects.filter(author=user)
+    return render(request, 'show_requests.html', {'user': user, 'lesson_requests': lesson_requests})
+
+'''allow users to change passwords'''
+@login_required
+def password(request):
+    current_user = request.user
+    if request.method == 'POST':
+        form = PasswordForm(data=request.POST)
+        if form.is_valid():
+            password = form.cleaned_data.get('password')
+            if check_password(password, current_user.password):
+                new_password = form.cleaned_data.get('new_password')
+                current_user.set_password(new_password)
+                current_user.save()
+                login(request, current_user)
+                messages.add_message(request, messages.SUCCESS, "Password updated!")
+                if isinstance(current_user, Admin):
+                    return redirect('admin_home')
+                else:
+                    return redirect('student_home')
+
+    form = PasswordForm()
+    return render(request, 'password.html', {'form': form})
 
 @login_required
+def profile(request):
+    current_user = request.user
+    if request.method == 'POST':
+        form = UserForm(instance=current_user, data=request.POST)
+        if form.is_valid():
+            messages.add_message(request, messages.SUCCESS, "Profile updated!")
+            form.save()
+            if isinstance(current_user, Admin):
+                return redirect('admin_home')
+            else:
+                return redirect('student_home')
+    else:
+        form = UserForm(instance=current_user)
+    return render(request, 'profile.html', {'form': form})
+
+@login_required
+@only_students
 def edit_requests(request, lesson_id):
     try:
         current_lesson = LessonRequest.objects.get(id=lesson_id)
@@ -192,11 +231,106 @@ def edit_requests(request, lesson_id):
             form = EditForm(instance=current_lesson)
         return render(request, 'edit_requests.html', {'form': form, 'lesson_id': lesson_id})
 
+@login_required
+@only_students
 def delete_requests(request, lesson_id):
     try:
         current_lesson = LessonRequest.objects.get(id=lesson_id)
     except ObjectDoesNotExist:
         return redirect('show_requests')
     else:
-        current_lesson.delete()
-        return redirect('show_requests')
+        user = request.user
+        if current_lesson.author != user:
+            return redirect('student_home')
+        else:
+            current_lesson.delete()
+            return redirect('show_requests')
+
+@login_required
+@only_students
+def show_invoices(request):
+    current_student = request.user
+    invoices = Invoice.objects.filter(student=current_student)
+    return render(request, 'invoices_list.html', {'invoices': invoices})
+
+
+
+@login_required
+@only_students
+def balance(request):
+    # first we need to get the student
+    current_student_id = request.user.id
+
+    student = Student.objects.get(id=current_student_id)
+
+    # then we retrieve all the lessons they have from the db
+    # invoices = Invoice.objects.filter(student_id=current_student_id)
+    invoices = student.unpaid_invoices
+    transfers = student.transfers
+    
+    # total money owed
+    total_due = 0
+    for invoice in invoices:
+        total_due += invoice.price
+
+    total_paid = 0
+    for transfer in transfers:
+        total_paid += transfer.invoice.price
+
+
+    return render(request, 'balance.html', {'invoices': invoices, 'transfers': transfers, 'total_paid':total_paid, 'total_due': total_due})
+
+
+@login_required
+@only_admins
+def all_student_balances(request):
+    all_students = Student.objects.all()
+    balances = {}
+    for student in all_students:
+        student_invoices = Invoice.objects.filter(student=student)
+        balance = 0
+        for invoice in student_invoices:
+            if(Transfer.objects.filter(invoice=invoice).count() == 0):
+                balance += invoice.price
+      
+        # Ensures only people with pending payments show up
+        if(balance != 0):
+            balances[student] = balance
+
+    return render(request, 'admin_payments.html', {'balances': balances})
+
+@login_required
+@only_admins
+def student_balance(request, student_id):
+    student = Student.objects.filter(id=student_id).first()
+    
+    transfer_list = student.transfers
+    invoice_list = student.unpaid_invoices
+    # student.invoices.exclude(id__in=transfer_list.values('invoice_id'))
+    
+    return render(request, 'admin_student_payments.html', {'invoices': invoice_list, 'transfers': transfer_list, 'student': student})
+
+@login_required
+@only_admins
+def approve_transaction(request, student_id, invoice_id):
+    if request.method == 'POST':
+        current_admin = request.user
+        invoice = Invoice.objects.filter(student_id=student_id).filter(invoice_number=invoice_id)
+        next_transfer_id = find_next_available_transfer_id()
+        transfer = Transfer.objects.create(date_received=timezone.now(), transfer_id=next_transfer_id, verifier=current_admin, invoice=invoice.first())
+        transfer.save()
+
+    return redirect('student_payments', student_id=student_id)
+
+@login_required
+@only_students
+def show_invoice_lessons(request, invoice_id):
+    """Shows the lessons associated with a given invoice"""
+    try:
+        current_invoice = Invoice.objects.get(id=invoice_id)
+    except ObjectDoesNotExist:
+        return redirect('show_invoices')
+    else:
+        lessons_to_display = current_invoice.lessons
+        return render(request, 'show_invoice_lessons.html', {'lessons': lessons_to_display, 'invoice':current_invoice})
+
